Java file structure modified
