+++ project/src/main/java/com/security/auth/data/VisualV1/V1Monthly.java
@@ -1,4 +1,4 @@
-package com.security.auth.data;
+package com.security.auth.data.VisualV1;
 
 import java.io.Serializable;
 
@@ -7,21 +7,23 @@ import javax.persistence.Id;
 import javax.persistence.IdClass;
 import javax.persistence.Table;
 
-
 @Entity
-@Table(name="v1Year")
-@IdClass(V1Yearly.class)
-public class V1Yearly implements Serializable {
+@Table(name="v1Monthly")
+@IdClass(V1Monthly.class)
+public class V1Monthly implements Serializable {
 
     @Id
     private int year;
+
+    @Id
+    private int month;
 
     @Id
     private String area;
 
     private double deg;
 
-    public int getYear() {
+    public double getYear() {
         return this.year;
     }
 
@@ -29,8 +31,12 @@ public class V1Yearly implements Serializable {
         return this.deg;
     }
 
+    public int getMonth() {
+        return this.month;
+    }
+
     public String getArea() {
-        return this.area;
+        return area;
     }
 
 }
