+++ project/src/main/java/com/security/auth/rest/VisualV1Rest/V1globalAnnualRest.java
@@ -1,4 +1,4 @@
-package com.security.auth.rest;
+package com.security.auth.rest.VisualV1Rest;
 
 import java.util.List;
 
@@ -6,8 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.security.auth.data.V1globalAnnual;
-import com.security.auth.service.V1globalAnnualService;
+import com.security.auth.data.VisualV1.V1globalAnnual;
+import com.security.auth.service.VisualV1Services.V1globalAnnualService;
 
 @RestController
 public class V1globalAnnualRest {
