+++ project/src/main/java/com/security/auth/service/VisualV1Services/V1globalAnnualService.java
@@ -1,12 +1,12 @@
-package com.security.auth.service;
+package com.security.auth.service.VisualV1Services;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.security.auth.data.V1globalAnnual;
-import com.security.auth.repo.V1globalAnnualRepo;
+import com.security.auth.data.VisualV1.V1globalAnnual;
+import com.security.auth.repo.VisualV1Repos.V1globalAnnualRepo;
 
 @Service
 public class V1globalAnnualService {
